refactor(benchmarks): rename Compass package
