+++ benchmarks/performance/compass/__init__.py
@@ -3,4 +3,3 @@
 RUN_SCHEMA = "compass.performance-run/1"
 SUITE_SCHEMA = "compass.performance-suite/1"
 WORKSPACE_SCHEMA = "compass.performance-workspace/1"
-
