+++ benchmarks/performance/compass/adapters.py
