+++ benchmarks/performance/compass/jsonstream.py
